refactor(researcher): Mover MetricsData a su propia entidad de dominio
- Creada nueva clase MetricsData en domain/entities/metrics_data.py
- Eliminada definición original de MetricsData de researcher.py
- Actualizada importación en researcher.py para usar la nueva ubicación
- Sin cambios funcionales, solo reorganización estructural
