+++ app/agents/researcher/domain/entities/__init__.py
@@ -1,5 +1,3 @@
-from .research_state import ResearchState
-from .section import Section
-from .search_query import SearchQuery
+from .metrics_data import MetricsData
 
-__all__ = ['ResearchState', 'Section', 'SearchQuery']
+__all__ = ['MetricsData']
