+++ app/agents/researcher/presentation/researcher.py
@@ -28,12 +28,9 @@ import sys
 
 from app.config.config import get_settings
 from app.services.tavilyService import tavily_search_async, deduplicate_and_format_sources
-from app.utils.state import ResearchState, SectionState, Section
-import os
-from pathlib import Path
-from contextlib import asynccontextmanager
+from app.utils.state import SectionState, Section
 from pydantic import BaseModel, Field
-from abc import ABC, abstractmethod
+
 
 # Configuración avanzada de logging
 logging.basicConfig(
@@ -46,33 +43,6 @@ logging.basicConfig(
 )
 logger = logging.getLogger(__name__)
 
-@dataclass
-class MetricsData:
-    """Structure to store performance metrics"""
-    start_time: float
-    end_time: float = 0
-    tokens_used: int = 0
-    api_calls: int = 0
-    errors: List[Dict] = None
-
-    def __post_init__(self):
-        self.errors = self.errors or []
-
-    @property
-    def duration(self) -> float:
-        """Calculate duration in seconds"""
-        if self.end_time:
-            return self.end_time - self.start_time
-        return time.time() - self.start_time
-
-    def to_dict(self) -> Dict:
-        """Convert metrics to dictionary"""
-        return {
-            'duration_seconds': self.duration,
-            'tokens_used': self.tokens_used,
-            'api_calls': self.api_calls,
-            'errors': self.errors
-        }
 
 def track_metrics(func):
     """Decorator to track function metrics"""
